add individualStrokeWeights to vector
