+++ lib/src/models/models.dart
@@ -15,6 +15,7 @@ export 'ellipse.dart';
 export 'export_setting.dart';
 export 'frame.dart';
 export 'group.dart';
+export 'individual_stroke_weights.dart';
 export 'instance.dart';
 export 'layout_align.dart';
 export 'layout_constraint.dart';
